fix: DashboardReadOnlyView accepts props, removes useParams
